[Attribute] Moving Attributes classes to a new Collector namespace for better organization.
